Class fixed
