+++ berichten/berichten.json
@@ -1,0 +1,1 @@
+{"firstName":"test2","lastName":"test3","message":"test1","postDate":"07-11-2022"}
